feat: Replace external cyrus-mcp-tools with inline SDK tools
- Implement inline cyrus tools using SDK's tool() callback feature from @anthropic-ai/claude-code v1.0.95
- Move LinearService and tool implementations to claude-runner package
- Update EdgeWorker to use inline tools instead of external MCP server
- Remove cyrus-mcp-tools package entirely
- Maintain full compatibility with agent session parent-child management
- Improve performance by eliminating stdio overhead and external process spawning

Benefits:
- Better performance with in-process execution
- Simpler deployment (tools bundled with claude-runner)
- Full TypeScript integration and type safety
- Direct access to application context and closures

All tests passing (55/55 in edge-worker)
