+++ packages/claude-runner/package.json
@@ -17,9 +17,13 @@
 	},
 	"dependencies": {
 		"@anthropic-ai/claude-code": "1.0.95",
-		"@anthropic-ai/sdk": "^0.60.0"
+		"@anthropic-ai/sdk": "^0.60.0",
+		"@linear/sdk": "^58.1.0",
+		"fs-extra": "^11.3.0",
+		"zod": "^4.1.4"
 	},
 	"devDependencies": {
+		"@types/fs-extra": "^11.0.4",
 		"@types/node": "^20.0.0",
 		"dotenv": "^16.5.0",
 		"typescript": "^5.3.3",
