+++ packages/claude-runner/src/index.ts
@@ -22,3 +22,5 @@ export type {
 	SDKSystemMessage,
 	SDKUserMessage,
 } from "./types.js";
+export { createCyrusToolsServer } from "./tools/cyrus-tools/index.js";
+export { LinearService } from "./tools/cyrus-tools/linear-service.js";
