--- packages/cyrus-mcp-tools/.editorconfig
@@ -1,10 +1,0 @@
-root = true
-
-[*]
-end_of_line = lf
-insert_final_newline = true
-charset = utf-8
-trim_trailing_whitespace = true
-
-[*.md]
-trim_trailing_whitespace = false
