--- packages/cyrus-mcp-tools/.env.example
@@ -1,15 +1,0 @@
-# Server Configuration
-PORT=3000
-NODE_ENV=development
-
-# Linear API
-LINEAR_API_URL=https://api.linear.app/graphql
-LINEAR_API_TOKEN=<YOUR_TOKEN>
-
-# Security
-CORS_ORIGINS=http://localhost:3000,https://your-deployed-domain.com
-RATE_LIMIT_WINDOW_MS=60000
-RATE_LIMIT_MAX=100
-
-# Logging
-LOG_LEVEL=info
