--- packages/cyrus-mcp-tools/.prettierrc
@@ -1,7 +1,0 @@
-{
-    "semi": true,
-    "trailingComma": "all",
-    "singleQuote": true,
-    "printWidth": 100,
-    "tabWidth": 2
-}
