--- packages/cyrus-mcp-tools/.vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-	"recommendations": [
-		"esbenp.prettier-vscode",
-		"dbaeumer.vscode-eslint",
-		"editorconfig.editorconfig"
-	]
-}
