--- packages/cyrus-mcp-tools/glama.json
@@ -1,4 +1,0 @@
-{
-	"$schema": "https://glama.ai/mcp/schemas/server.json",
-	"maintainers": ["beautyfree"]
-}
