--- packages/cyrus-mcp-tools/jest.config.js
@@ -1,22 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-export default {
-	preset: "ts-jest/presets/default-esm",
-	testEnvironment: "node",
-	roots: ["<rootDir>/src"],
-	testMatch: ["**/__tests__/**/*.ts?(x)", "**/?(*.)+(spec|test).ts?(x)"],
-	collectCoverageFrom: ["src/**/*.{ts,tsx}", "!src/**/*.d.ts", "!src/index.ts"],
-	coverageDirectory: "coverage",
-	transform: {
-		"^.+\\.tsx?$": [
-			"ts-jest",
-			{
-				tsconfig: "tsconfig.json",
-				useESM: true,
-			},
-		],
-	},
-	extensionsToTreatAsEsm: [".ts"],
-	moduleNameMapper: {
-		"^(\\.{1,2}/.*)\\.js$": "$1",
-	},
-};
