--- packages/cyrus-mcp-tools/nodemon.json
@@ -1,6 +1,0 @@
-{
-	"watch": ["src"],
-	"ext": "ts,json",
-	"ignore": ["src/**/*.spec.ts"],
-	"exec": "node --loader ts-node/esm src/index.ts"
-}
