--- packages/cyrus-mcp-tools/package.json
@@ -1,84 +1,0 @@
-{
-	"name": "cyrus-mcp-tools",
-	"version": "0.3.0",
-	"description": "MCP tools for Cyrus - Linear file uploads, agent session creation, and other utilities.",
-	"main": "dist/index.js",
-	"type": "module",
-	"bin": {
-		"cyrus-mcp-tools": "./dist/index.js"
-	},
-	"directories": {
-		"doc": "docs"
-	},
-	"scripts": {
-		"start": "node dist/index.js",
-		"inspect": "npx @modelcontextprotocol/inspector node dist/index.js",
-		"dev": "nodemon --exec node --loader ts-node/esm src/index.ts",
-		"build": "tsc",
-		"lint": "eslint . --ext .ts",
-		"format": "prettier --write \"src/**/*.ts\"",
-		"test": "jest --passWithNoTests",
-		"test:watch": "jest --watch",
-		"prepare": "npm run build",
-		"postinstall": "node -e \"try { require('fs').chmodSync('./dist/index.js', '755') } catch (e) {}\""
-	},
-	"smithery": {
-		"name": "cyrus-mcp-tools",
-		"displayName": "Cyrus MCP Tools",
-		"description": "MCP tools for Cyrus including Linear file uploads and agent session management",
-		"tools": [
-			"linear_upload_file",
-			"linear_agent_session_create",
-			"linear_agent_give_feedback"
-		]
-	},
-	"keywords": [
-		"mcp",
-		"linear",
-		"cursor",
-		"claude",
-		"ai",
-		"model-context-protocol",
-		"project-management",
-		"smithery"
-	],
-	"author": "Cyrus AI",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/ceedaragents/cyrus"
-	},
-	"dependencies": {
-		"@linear/sdk": "^38.0.0",
-		"@modelcontextprotocol/sdk": "^1.6.0",
-		"@types/cors": "^2.8.17",
-		"@types/express": "^5.0.0",
-		"cors": "^2.8.5",
-		"dotenv": "^16.4.7",
-		"express": "^4.21.2",
-		"graphql": "^16.10.0",
-		"graphql-request": "^7.1.2",
-		"helmet": "^8.0.0",
-		"yargs": "^17.7.2",
-		"zod": "^3.24.2"
-	},
-	"devDependencies": {
-		"@types/jest": "^29.5.14",
-		"@types/node": "^22.13.5",
-		"@types/yargs": "^17.0.33",
-		"eslint": "^9.21.0",
-		"eslint-config-prettier": "^10.0.2",
-		"jest": "^29.7.0",
-		"nodemon": "^3.1.9",
-		"prettier": "^3.5.2",
-		"ts-jest": "^29.2.6",
-		"ts-node": "^10.9.2",
-		"typescript": "^5.7.3"
-	},
-	"engines": {
-		"node": ">=20.0.0"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
-}
