--- packages/cyrus-mcp-tools/smithery.yaml
@@ -1,21 +1,0 @@
-# Smithery configuration file: https://smithery.ai/docs/config#smitheryyaml
-
-startCommand:
-  type: stdio
-  configSchema:
-    # JSON Schema defining the configuration options for the MCP
-    type: object
-    required:
-      - token
-    properties:
-      token:
-        type: string
-        description: Your Linear API key. Required for Linear API access.
-  commandFunction: |
-    (config) => ({
-      command: 'node',
-      args: ['dist/index.js', '--token', config.token],
-      env: {
-        LINEAR_API_TOKEN: config.token
-      }
-    })
