--- packages/cyrus-mcp-tools/src/types.ts
@@ -1,17 +1,0 @@
-/**
- * Interface for MCP Tool Definition in our format
- */
-export interface MCPToolDefinition {
-	name: string;
-	description: string;
-	input_schema: {
-		type: string;
-		properties: Record<string, any>;
-		required?: string[];
-	};
-	output_schema: {
-		type: string;
-		properties?: Record<string, any>;
-		items?: any;
-	};
-}
