+++ packages/edge-worker/test/EdgeWorker.dynamic-tools.test.ts
@@ -185,7 +185,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(debuggerTools).toEqual([
 				...getReadOnlyTools(),
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 
 			// Test builder prompt with custom array
@@ -195,7 +195,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 				"Edit",
 				"Task",
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 
 			// Test scoper prompt with safe preset
@@ -203,7 +203,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(scoperTools).toEqual([
 				...getSafeTools(),
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 
@@ -235,7 +235,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(debuggerTools).toEqual([
 				...getAllTools(),
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 
 			// Test builder prompt with global safe preset
@@ -243,7 +243,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(builderTools).toEqual([
 				...getSafeTools(),
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 
 			// Test scoper prompt with global custom array
@@ -252,7 +252,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 				"Read",
 				"WebFetch",
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 
@@ -269,7 +269,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 				"Read",
 				"Write",
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 
@@ -287,7 +287,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 				"Write",
 				"Edit",
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 
@@ -308,7 +308,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(tools).toEqual([
 				...getSafeTools(),
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 
@@ -322,7 +322,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			const tools = buildAllowedTools(repository);
 
 			// Should deduplicate Linear MCP tools
-			expect(tools).toEqual(["Read", "mcp__linear", "mcp__cyrus-mcp-tools"]);
+			expect(tools).toEqual(["Read", "mcp__linear", "mcp__cyrus-tools"]);
 			expect(tools.filter((t) => t === "mcp__linear")).toHaveLength(1);
 		});
 
@@ -347,7 +347,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 				"Write",
 				"Edit",
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 
 			// New format should work as expected
@@ -355,7 +355,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(builderTools).toEqual([
 				...getSafeTools(),
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 
@@ -376,7 +376,7 @@ describe("EdgeWorker - Dynamic Tools Configuration", () => {
 			expect(tools).toEqual([
 				"CustomTool",
 				"mcp__linear",
-				"mcp__cyrus-mcp-tools",
+				"mcp__cyrus-tools",
 			]);
 		});
 	});
